Remove unused npm deps, use prod instead of dev scripts
We also stopped using bind mounts for now,
since it resolves the problem of having to install node_modules on the host.
This is caused by node_modules folder being overwritten when the bind mount is applied.
One solution to this, and something we might explore,
is having bind mounts only for the databases,
since they are the only ones that need persistence.